% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_diag_correction_cpp}
\alias{compute_diag_correction_cpp}
\title{Compute diagonal correction for V update (C++ version)}
\usage{
compute_diag_correction_cpp(U)
}
\arguments{
\item{U}{Factor matrix}
}
\value{
R x R diagonal correction matrix
}
\description{
Computes the diagonal correction term sum_i U_sq(i,r1) * U_sq(i,r2)
used in updateV and updateVm functions.
}
\keyword{internal}
