% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_L_matrix_cpp}
\alias{compute_L_matrix_cpp}
\title{Compute L matrix for V update (C++ version)}
\usage{
compute_L_matrix_cpp(Z, U, T_dim, R)
}
\arguments{
\item{Z}{3D array (N x N x T), already zeroed below diagonal}

\item{U}{Factor matrix (N x R)}

\item{T_dim}{Number of time points}

\item{R}{Number of latent dimensions}
}
\value{
T_dim x R matrix L
}
\description{
Computes L(t, r) = sum over i less than j of Z(i,j,t) * U(i,r) * U(j,r)
This is a key bottleneck in updateV.
}
\keyword{internal}
