% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GramSchmidt_cpp}
\alias{GramSchmidt_cpp}
\title{Gram-Schmidt Orthogonalization (C++ version)}
\usage{
GramSchmidt_cpp(U)
}
\arguments{
\item{U}{Input matrix to orthogonalize}
}
\value{
Orthonormalized matrix
}
\description{
Gram-Schmidt Orthogonalization (C++ version)
}
\keyword{internal}
