% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{M_U_cpp}
\alias{M_U_cpp}
\title{Create 3D Array from Factor Matrices (C++ version)}
\usage{
M_U_cpp(U1, U2, V)
}
\arguments{
\item{U1}{First factor matrix (N1 x R)}

\item{U2}{Second factor matrix (N2 x R), typically same as U1 for undirected networks}

\item{V}{Third factor matrix (T x R)}
}
\value{
3D array of dimension (N1 x N2 x T)
}
\description{
Computes M(i,j,t) = sum_r U1(i,r) * U2(j,r) * V(t,r)
This is the core tensor reconstruction operation used in NetworkChange.
}
\keyword{internal}
