% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{M_U_list_cpp}
\alias{M_U_list_cpp}
\title{Create 3D Array from List of Factor Matrices (C++ version)}
\usage{
M_U_list_cpp(U_list)
}
\arguments{
\item{U_list}{List of factor matrices}
}
\value{
Array with dimensions corresponding to the rows of each factor matrix
}
\description{
General version that takes a list of factor matrices.
For NetworkChange, typically called with list(U, U, V) for undirected networks.
}
\keyword{internal}
