\name{multiplot}
\alias{multiplot}
\title{
 Printing multiple ggplots in oone file
}
\description{
 Print multiple ggplots in one file. Slightly modified for packaging from the original version in the web. 
}
\usage{
multiplot(..., plotlist = NULL, cols = 1, layout = NULL)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{\dots}{
  A list of ggplot objects separated by commas.
}
  \item{plotlist}{
  A list of ggplot objects.
  
}
  \item{cols}{
  The number of columns.
}
  \item{layout}{
  A matrix specifying the layout. If present, 'cols' is ignored.
}
}
\author{
	\url{http://www.cookbook-r.com/Graphs/Multiple_graphs_on_one_page_(ggplot2)/}
}
