% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkChangeCVtest.R
\name{NetworkChangeCVtest}
\alias{NetworkChangeCVtest}
\title{k-fold Cross-Validation for Network Change point models}
\usage{
NetworkChangeCVtest(Y, R = 2, k.fold = 2, mcmc = 100, burnin = 100,
  thin = 1, verbose = 0, LOO = FALSE, LOO.bound = 2, break.upper = 3,
  Waic = FALSE, v0 = NULL, v1 = NULL)
}
\arguments{
\item{Y}{Reponse tensor}

\item{R}{Dimension of latent space. The default is 2.}

\item{k.fold}{The number of folds that should be used for cross-validation.}

\item{mcmc}{The number of MCMC iterations after burnin.}

\item{burnin}{The number of burn-in iterations for the sampler.}

\item{thin}{The thinning interval used in the simulation.  The number of
MCMC iterations must be divisible by this value.}

\item{verbose}{A switch which determines whether or not the progress of the
sampler is printed to the screen.  If \code{verbose} is greater than 0 the
iteration number, the \eqn{\beta} vector, and the error variance are
printed to the screen every \code{verbose}th iteration.}

\item{LOO}{If \code{TRUE}, do the leave-one-out cross-validation.}

\item{LOO.bound}{The boundary of LOO test. By default, \code{LOO.bound=2}, meaning that first and last 2 observations are dropped
in the leave-one-out cross-validation.}

\item{break.upper}{The upper limit of break number for cross-validation test.}
}
\value{
Mean squared error of the k-fold cross validation test
}
\description{
NetworkChangeCVtest implements a suite of k-fold cross-validation test for network change point models
}
\examples{

   \dontrun{
   set.seed(1973)

   ## generate an array with three constant blocks
   set.seed(11173)
   N <- 10
   Y <- MakeBlockNetworkChange(n=N, break.point = .5, base.prob=.2, block.prob=.5,
        shape=1, T=20, type ="split")
   G <- 10 ## Small mcmc scans to save time
   MSE <- NetworkChangeCVtest(Y, k.fold = 2)

   ## find mean of MSE
   apply(MSE, 2, mean)

   }

}
\references{
Jong Hee Park and Yunkyun Sohn. 2017. "Detecting Structural Change
in Network Time Series Data using Bayesian Inference." Working Paper.
}
\seealso{
\code{\link{NetworkChange}}
}
