% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rmn_cpp}
\alias{rmn_cpp}
\title{Matrix Normal Random Sampling (C++ version)}
\usage{
rmn_cpp(M, Srow, Scol)
}
\arguments{
\item{M}{Mean matrix}

\item{Srow}{Row covariance matrix}

\item{Scol}{Column covariance matrix}
}
\value{
Random matrix from Matrix Normal(M, Srow, Scol)
}
\description{
Matrix Normal Random Sampling (C++ version)
}
\keyword{internal}
