% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rMVNorm_cpp}
\alias{rMVNorm_cpp}
\title{Multivariate Normal Random Sampling (C++ version)}
\usage{
rMVNorm_cpp(n, mu, Sigma)
}
\arguments{
\item{n}{Number of samples}

\item{mu}{Mean vector}

\item{Sigma}{Covariance matrix}
}
\value{
Matrix of n samples (rows) from MVN(mu, Sigma)
}
\description{
Multivariate Normal Random Sampling (C++ version)
}
\keyword{internal}
