% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rsmn_cpp}
\alias{rsmn_cpp}
\title{Standard Normal Matrix Generator (C++ version)}
\usage{
rsmn_cpp(m, n)
}
\arguments{
\item{m}{Number of rows}

\item{n}{Number of columns}
}
\value{
m x n matrix of standard normal random values
}
\description{
Standard Normal Matrix Generator (C++ version)
}
\keyword{internal}
