\name{startS}
\alias{startS}
\title{
Sample a starting value of hidden states
}
\description{
Sample a starting value of hidden states
}
\usage{
startS(Z, Time, m, initial.U, V, s2, R)
}
\arguments{
  \item{Z}{
  Degree-corrected network array data.
}
  \item{Time}{
 The length of time. 
}
  \item{m}{
  The number of breaks
}
  \item{initial.U}{
  Initialized U matrix.
}
  \item{V}{
  Initialized V matrix.
}
  \item{s2}{
  The variance of error.
}
  \item{R}{
  The dimensionality of latent space
}
}
