\name{startUV}
\alias{startUV}
\title{
Starting values of U and V
}
\description{
Initialize starting values of U and V

}
\usage{
startUV(Z, R, K)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{Z}{
  Degree-corrected network array data.
  }
  \item{R}{
  The dimensionality of latent space.
}
  \item{K}{
  The dimensionality of Z.
}
}
