% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ULUstateSample.r
\name{ULUstateSample}
\alias{ULUstateSample}
\title{Hidden State Sampler}
\usage{
ULUstateSample(m, s, ZMUt, s2, P, SOS.random)
}
\arguments{
\item{m}{The number of break}

\item{s}{Latent state vector}

\item{ZMUt}{Z - MU}

\item{s2}{error variance}

\item{P}{Transition matrix}

\item{SOS.random}{single observation state random perturbation}
}
\value{
A list of a state vector, state probabilities, and SOS.random.
}
\description{
Sample hidden states from hidden Markov multilinear model.
Uses vectorized operations for improved performance.
}
