% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ULUstateSample_cpp}
\alias{ULUstateSample_cpp}
\title{Hidden State Sampler (C++ version)}
\usage{
ULUstateSample_cpp(m, s, ZMUt, s2, P, SOS_random)
}
\arguments{
\item{m}{Number of breaks}

\item{s}{Current state vector (not used, kept for interface)}

\item{ZMUt}{List of (Z - MU) matrices for each state}

\item{s2}{Vector of error variances for each state}

\item{P}{Transition probability matrix}

\item{SOS_random}{Whether to apply single observation state randomization}
}
\value{
List containing: s (state vector), ps (state probabilities), SOS (flag)
}
\description{
Sample hidden states from hidden Markov multilinear model using
forward filtering backward sampling algorithm.
}
\keyword{internal}
