% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Unormal_cpp}
\alias{Unormal_cpp}
\title{Column Normalization (C++ version)}
\usage{
Unormal_cpp(U1)
}
\arguments{
\item{U1}{Input matrix}
}
\value{
Matrix with normalized columns
}
\description{
Normalize columns of a matrix to have unit Euclidean norm.
}
\keyword{internal}
