\name{updateb}
\alias{updateb}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Update time-constant beta.
}
\description{
Update time-constant beta.
}
\usage{
updateb(Z, MU, s2, XtX, b0, B0)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{Z}{
Degree-corrected network array data.
}
  \item{MU}{
Mean arrary.
}
  \item{s2}{
Residual variance. 
}
  \item{XtX}{
X^T%*%X. 
}
  \item{b0}{
Prior mean of beta
}
  \item{B0}{
Prior variance of beta
}
}
