\name{updateP}
\alias{updateP}
\title{
Update P.
}
\description{
Update P.}
\usage{
updateP(s, ns, P, A0)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{s}{
Latent states.
}
  \item{ns}{
Number of unique hidden states.
}
  \item{P}{
Previous P. 
}
  \item{A0}{
Prior of P. 
}
}
