\name{updates2m}
\alias{updates2m}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Update regime-changing regression variances. 
}
\description{
Update regime-changing regression variances. 
}
\usage{
updates2m(ns, Zm, MU, c0, d0, Km)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{ns}{
  The number of hidden regimes.
}
  \item{Zm}{
  The regime-specific holder of Z - beta.
}
  \item{MU}{
  The mean array. 
}
  \item{c0}{
  Scalar shape parameter.
}
  \item{d0}{
  Scalar scale parameter
}
  \item{Km}{
  Regime-specific dimensions
}
}
