% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{updates2m_cpp}
\alias{updates2m_cpp}
\title{Update Regime-specific Variance (C++ version)}
\usage{
updates2m_cpp(ns, Zm, MU, c0, d0, Km)
}
\arguments{
\item{ns}{Number of hidden states}

\item{Zm}{List of regime-specific (Z - beta) arrays}

\item{MU}{List of mean arrays}

\item{c0}{Shape parameter prior}

\item{d0}{Scale parameter prior}

\item{Km}{List of regime-specific dimensions}
}
\value{
Vector of regime-specific variances
}
\description{
Updates regime-specific variance parameters using inverse gamma sampling.
}
\keyword{internal}
