\name{updateU}
\alias{updateU}
\title{
Update time-constant latent node positions. 
}
\description{
Update time-constant latent node positions. 
}
\usage{
updateU(K, U, V, R, Zb, s2, eU, iVU)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{K}{
 The dimension of Z.
}
  \item{U}{
  Latent node positions.
}
  \item{V}{
  Layer-specific network generation rule.
}
  \item{R}{
  The dimensionality of latent space
}
  \item{Zb}{
  Z - beta
}
  \item{s2}{
  The variance of error
}
  \item{eU}{
  The mean of U
}
  \item{iVU}{
 The variance of U
}
}
