% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{updateU_cpp}
\alias{updateU_cpp}
\title{Update Time-constant Latent Node Positions (C++ version)}
\usage{
updateU_cpp(K, U, V, R, Zb, s2, eU, iVU)
}
\arguments{
\item{K}{Dimensions of Z (N, N, T)}

\item{U}{Current latent node positions (N x R)}

\item{V}{Layer-specific network generation rule (T x R)}

\item{R}{Dimension of latent space}

\item{Zb}{3D array (Z - beta)}

\item{s2}{Error variance (scalar)}

\item{eU}{Prior mean vector for U}

\item{iVU}{Prior precision matrix for U}
}
\value{
Updated U matrix (N x R)
}
\description{
Updates U using optimized matrix operations with pre-computed invariants.
}
\keyword{internal}
