\name{update.V}
\alias{updateV}
\title{
Update layer specific network generation rules
}
\description{
Update layer specific network generation rules.
}
\usage{
updateV(Zb, U, R, K, s2, eV, iVV, UTA)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{Zb}{
 Z - beta.
}
  \item{U}{
 The latent node positions.
}
  \item{R}{
The dimension of latent space.
}
  \item{K}{
The dimension of Z.
}
  \item{s2}{
The variance of error.
}
  \item{eV}{
The mean of V.
}
  \item{iVV}{
The variance of V.
}
  \item{UTA}{
Indicator of upper triangular array 
}
}
