% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaicCompare.R
\name{WaicCompare}
\alias{WaicCompare}
\title{Compare WAIC}
\usage{
WaicCompare(outlist)
}
\arguments{
\item{outlist}{List of NetworkChange objects}
}
\value{
out

A matrix of log marginal likelihoods.
}
\description{
Compare Widely Applicable Information Criterion
}
\references{
Sumio Watanabe. 2010. "Asymptotic equivalence of Bayes cross validation and widely
applicable information criterion in singular learning theory."
\emph{Journal of Machine Learning Research}. 11: 3571-3594.

Andrew Gelman, Jessica Hwang, and Aki Vehtari. 2014. "Understanding predictive information
criteria for Bayesian models." \emph{Statistics and Computing}. 24(6):997-1016.
}
\seealso{
\code{\link{MarginalCompare}}
}
