% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BreakDiagnostic.R
\name{BreakDiagnostic}
\alias{BreakDiagnostic}
\title{Detect a break number using different metrics}
\usage{
BreakDiagnostic(Y, R = 2, mcmc = 100, burnin = 100, verbose = 100,
  thin = 1, UL.Normal = "Orthonormal", v0 = NULL, v1 = NULL,
  break.upper = 3)
}
\arguments{
\item{Y}{Reponse tensor}

\item{R}{Dimension of latent space. The default is 2.}

\item{mcmc}{The number of MCMC iterations after burnin.}

\item{burnin}{The number of burn-in iterations for the sampler.}

\item{verbose}{A switch which determines whether or not the progress of the
sampler is printed to the screen.  If \code{verbose} is greater than 0 the
iteration number, the \eqn{\beta} vector, and the error variance are
printed to the screen every \code{verbose}th iteration.}

\item{thin}{The thinning interval used in the simulation.  The number of
MCMC iterations must be divisible by this value.}

\item{UL.Normal}{Transformation of sampled U. Users can choose "NULL", "Normal" or "Orthonormal."
"NULL" is no normalization. "Normal" is the standard normalization.
"Orthonormal" is the Gram-Schmidt orthgonalization. Default is "NULL."}

\item{v0}{\eqn{v_0/2} is the shape parameter for the inverse
Gamma prior on variance parameters for V.
If \code{v0 = NULL}, a value is computed from a test run of \code{NetworkStatic}.}

\item{v1}{\eqn{v_1/2} is the scale parameter for the
inverse Gamma prior on variance parameters for V.
If \code{v1 = NULL}, a value is computed from a test run of \code{NetworkStatic}.}

\item{break.upper}{Upper threshold for break number detection.
The default is \code{break.upper = 3}.}

\item{degree.normal}{A null model for degree correction. Users can choose "NULL", "eigen" or "Lsym."
"NULL" is no degree correction. "eigen" is a principal eigen-matrix consisting of
the first eigenvalue and the corresponding eigenvector. "
Lsym" is a modularity matrix. Default is "eigen."}
}
\description{
Detect a break number using different metrics
}
\examples{
   \dontrun{
   set.seed(1973)
   ## Generate an array (30 by 30 by 40) with block transitions
   ## One break model (splitting)
   Y <- MakeBlockNetworkChange(n=10, T=40, type ="split")

   ## Fit multiple models for break number detection using Bayesian model comparison
   detect <- BreakDiagnostic(Y, R=2)
   
   ## Look at the graph
   detect[[1]]; print(detect[[2]])


   ## Two break model
   Y <- MakeBlockNetworkChange(n=10, T=40, type ="split-merge")
}


}
