% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarginalCompare.R
\name{MarginalCompare}
\alias{MarginalCompare}
\title{Compare Log Marginal Likelihood}
\usage{
MarginalCompare(outlist)
}
\arguments{
\item{outlist}{List of NetworkChange objects}
}
\value{
A matrix of log marginal likelihoods.
}
\description{
Compare Log Marginal Likelihood
}
\references{
Siddhartha Chib. 1995. ``Marginal Likelihood from the Gibbs Output.''
\emph{Journal of the American Statistical Association}. 90: 1313-1321.
}
\seealso{
\code{\link{WaicCompare}}
}
