% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot.R
\name{multiplot}
\alias{multiplot}
\title{Printing multiple ggplots in oone file}
\usage{
multiplot(..., plotlist = NULL, cols = 1, layout = NULL)
}
\arguments{
\item{...}{A list of ggplot objects separated by commas.}

\item{plotlist}{A list of ggplot objects}

\item{cols}{The number of columns.}

\item{layout}{A matrix specifying the layout. If present, 'cols' is ignored.}
}
\value{
A plot object
}
\description{
Print multiple ggplots in one file. Slightly modified for packaging from the original version in the web.
}
\author{
\url{http://www.cookbook-r.com/Graphs/Multiple_graphs_on_one_page_(ggplot2)/}
}
