% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startUV.R
\name{startUV}
\alias{startUV}
\title{Starting values of U and V}
\usage{
startUV(Z, R, K)
}
\arguments{
\item{Z}{Degree-corrected network array data.}

\item{R}{The dimensionality of latent space.}

\item{K}{The dimensionality of Z.}
}
\value{
A list of U and V
}
\description{
Initialize starting values of U and V
}
