% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateb.R
\name{updateb}
\alias{updateb}
\title{Update time-constant regression parameters}
\usage{
updateb(Z, MU, s2, XtX, b0, B0)
}
\arguments{
\item{Z}{Degree corrected response tensor}

\item{MU}{Mean array}

\item{s2}{Error variance}

\item{XtX}{X'X}

\item{b0}{Prior mean of beta}

\item{B0}{Prior variance of beta}
}
\value{
A vector of regression parameters
}
\description{
Update time-constant regression parameters
}
