% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updatebm.R
\name{updatebm}
\alias{updatebm}
\title{Update regime-changing regression parameters}
\usage{
updatebm(ns, K, s, s2, B0, p, ZU)
}
\arguments{
\item{ns}{The number of hidden states}

\item{K}{The dimensionality of Z}

\item{s}{Latent state vector}

\item{s2}{The variance of error}

\item{B0}{The prior variance of beta}

\item{p}{The rank of X}

\item{ZU}{Z - ULU}
}
\value{
A vector of regime-changing regression parameters
}
\description{
Update regime-changing beta
}
