% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateP.R
\name{updateP}
\alias{updateP}
\title{Update transition matrix}
\usage{
updateP(s, ns, P, A0)
}
\arguments{
\item{s}{Latent state vector}

\item{ns}{The number of hidden states}

\item{P}{Transition matrix}

\item{A0}{Prior of transition matrix}
}
\value{
A transtion matrix
}
\description{
Update transition matrix
}
