% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateS.R
\name{updateS}
\alias{updateS}
\title{Update latent states}
\usage{
updateS(iter, s, V, m, Zb, Zt, Time, MU.state, P, s2, N.upper.tri,
  random.perturb)
}
\arguments{
\item{iter}{iteration number}

\item{s}{the most recent latent states}

\item{V}{Network generation rules}

\item{m}{The number of breaks}

\item{Zb}{Z - b}

\item{Zt}{Z stacked by time}

\item{Time}{The length of time}

\item{MU.state}{UVU for each state}

\item{P}{Transition matrix}

\item{s2}{error variance}

\item{N.upper.tri}{The number of upper triangular elements}

\item{random.perturb}{If \code{random.perturb} = TRUE and a single state observation is found,
the latent state is randomly selected by equal weights.}
}
\value{
A list of vectors containing latent states and their probabilities
}
\description{
Update latent states
}
