% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updates2m.R
\name{updates2m}
\alias{updates2m}
\title{Update regime-specific variance}
\usage{
updates2m(ns, Zm, MU, c0, d0, Km)
}
\arguments{
\item{ns}{The number of hidden states}

\item{Zm}{The regime-specific holder of Z - beta}

\item{MU}{The mean array.}

\item{c0}{Scalar shape parameter}

\item{d0}{Scalar scale parameter}

\item{Km}{Regime-specific dimensions}
}
\value{
A scalar for a regime-specific variance
}
\description{
Update regime-specific variance parameter
}
