% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateU.R
\name{updateU}
\alias{updateU}
\title{Update time-constant latent node positions}
\usage{
updateU(K, U, V, R, Zb, s2, eU, iVU)
}
\arguments{
\item{K}{The dimensionality of Z}

\item{U}{The most recent draw of latent node positions}

\item{V}{Layer-specific network generation rule}

\item{R}{The dimensionality of latent space}

\item{Zb}{Z - beta}

\item{s2}{error variance}

\item{eU}{The mean of U}

\item{iVU}{The variance of U}
}
\value{
A matrix of time-constant latent node positions
}
\description{
Update time-constant latent node positions
}
